requirements.txt
# Python package dependencies for MCP Database Report Generator
pandas>=1.5.0
pyodbc>=4.0.0
xlsxwriter>=3.0.0
python-dotenv>=1.0.0
